Move output::log to output::build_log
It's repetitive to have `build_log::BuildLog` but it's otherwise odd that there are logging things that aren't in `mod log`.

Now this import `use commons::output::build_log::*;` will give you everything you need to use `BuildLog`.
