+++ buildpacks/ruby/src/gem_list.rs
@@ -1,9 +1,8 @@
 #[allow(clippy::wildcard_imports)]
-use commons::output::section_log::*;
+use commons::output::{fmt, section_log::*};
 
 use commons::fun_run::{CmdError, CommandWithName};
 use commons::gem_version::GemVersion;
-use commons::output::fmt;
 use core::str::FromStr;
 use regex::Regex;
 use std::collections::HashMap;
