+++ buildpacks/ruby/src/steps/rake_assets_install.rs
@@ -1,9 +1,11 @@
+#[allow(clippy::wildcard_imports)]
+use commons::output::{fmt, section_log::*};
+
 use crate::rake_task_detect::RakeDetect;
 use crate::RubyBuildpack;
 use crate::RubyBuildpackError;
 use commons::cache::{mib, AppCacheCollection, CacheConfig, KeepPath};
 use commons::fun_run::{self, CmdError, CommandWithName};
-use commons::output::{fmt, section_log::*};
 use libcnb::build::BuildContext;
 use libcnb::Env;
 use std::process::Command;
