+++ commons/src/output/mod.rs
@@ -1,6 +1,6 @@
 mod background_timer;
+pub mod build_log;
 pub mod fmt;
 pub mod interface;
-pub mod log;
 pub mod section_log;
 mod util;
